Add files via upload
There is one solution here that must be credited to an external source (stratascratch on youtube). The amazon marketing campaign question as the logic there was difficult and i missed 2 key scenarios.
I kinda have an idea what was now done but the logic doesn't feel natural. The technical skills used on the marketing question is not strange to me but the logic however did a one on me. 
I am most proud of the fb_messages question as i leveraged excel to get a deeper look into sql as i was feeling overwhelmed learning a new skill in a matter of weeks and it worked. That is the lesson to take from that.