Update README.md
# sql
an sql repo made to test git integration with azure data studio
Credit for the solution to the amz_marketing campaign stratq question must go to Stratascratch (from youtube). The technical skills like concat and ranking over partition were no stranger to me but the logic in the answer did a number on me. To my credit i was not connected to the data as i was in the fb_messages questions and i was also super focused on the interview setting.
However this is the lesson, start slow linking sql to what you already know excel, R, python and then you will see your speed increase. It took me some time to ease into the fb_messages q but once i did then i was able to leverage all my tools to help me solve it. I will be solving more but trust me i was proud of it because learning sql in the time frame i used was overwhelming and i got lost in it and forgot the other things i knew.
Thanks stratascratch for amazing work and kudos to myself on solving my own strata Q.